flask/requirements.txt
flask
flask_cors
pandas
openai
python-dotenv
gunicorn
requests